% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7.get_ord_axis_labels.R
\name{get_ord_axis_labels}
\alias{get_ord_axis_labels}
\title{Internal function to get axis labels for vegan ordination objects}
\usage{
get_ord_axis_labels(ord)
}
\arguments{
\item{ord}{A vegan ordination object.}
}
\value{
A list with one element \code{label}, a character vector for axis labels.
#' @examples
library(vegan)
data(dune)
data(dune.env)
ord_pca <- rda(dune)
get_ord_axis_labels(ord_pca)
}
\description{
This function infers the ordination method used and returns proper axis labels
including percentage of explained variance when applicable.
}
